[TASK] Existing Backend group required
A new extension setting (TYPO3BEGroupExist) is introduced for the
Backend login that checks if the user is a member in at least one
existing Backend group. If not, the login will not be valid.

This option was already available for the Frontend login and only
the setting for enabling it for the Backend is added.

Change-Id: Ib7b7a912be252ce60283b677b948d140fdbc07d0
Resolves: #59094
Reviewed-on: https://review.typo3.org/30382
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers