[BUGFIX] No local groups shown for backend/frontend users in TCA
This bug is related to a faulty MySQL driver when running an old enterprise
version of RedHat. See https://bugzilla.redhat.com/show_bug.cgi?id=349121
for details.

Although the database table does not allow NULL for the tx_igldapssoauth_dn
column, the MySQL driver for PHP interprets empty string as a NULL value,
which in turn prevents the TCA conditions from being fulfilled.

The additional "IS NULL" / "IS NOT NULL" conditions ensure that this rare
case does not negatively impact this extension and let it deal and recover
from this bug silently.

Fixes: #66189
Related: #66055
Change-Id: I91c3a14eff0dae03d44a86f17e0a89b5ceedae97
Reviewed-on: http://review.typo3.org/38376
Reviewed-by: Xavier Perseguers <xavier@typo3.org>
Tested-by: Xavier Perseguers <xavier@typo3.org>