[BUGFIX] Pass LDAP arguments during AJAX import
If a user is imported manually using the backend module,
then the import doesn’t use the configured attributes for the LDAP server.
The resulting search request therefore may contain incomplete
data and the import will fail.

To fix this we pass the LDAP attributes from the mapping
given in the LDAP configuration.

Refs #15