[TASK] Drop support for checking extension configuration
This was useful for former versions of TYPO3 when the security level
for how to handle hashed (or not) passwords could be wrongly configured
to be then used by this LDAP extension. But this is not a problem
anymore nowadays and in addition, rendering is absolutely not handy
anymore since TYPO3 v9 as extension configuration is rendered via
AJAX and does not properly update after saving.