[BUGFIX] Prevent new user being disabled with first login
Since TYPO3 v8, new users are disabled by default. Since this extension
starts by creating a default record for new users and then updates this
record with the LDAP mapping, the query fetching the newly-created row
needs to remove the filter on disabled records. To be on the safe side,
all restrictions are removed altogether.