[TASK] Allow SSO with AD-Forest
When using this extension in a large AD-Forest, SSO may authenticate
users as the wrong one since domain is stripped prior to authenticating.

Example:

Peter is visiting the page as RemoteUser maier@xyz.example.com, but may
get authenticated as Kathrin maier@abc.example.com if the LDAP configuration
for abc.example.com is priorized before xyz.example.com.

When switching the priority of these two, Kathrin and Peter would get
authenticated as Peter.

This fixes the issue by including the DN into the SSO algorithm.

Resolves: #73