[TASK] Enhance PSR-14 user added/updated events with extraData
In an attempt to get rid of the old hook for processing custom
extra data, the two events are extended with the custom extra
data from the mapping.

The hook is not actually deprecated yet as version 4.0.0 is
already out. This is a quick backward-compatible addition so
that hook may already be migrated in version 4.0.1 while getting
officially deprecated in version 4.1.0.